PixelAutoClicker/requirements.txt
pyautogui
keyboard
Pillow
customtkinter